refactor: share docs slug normalization
